evil-base-treediff: new strategy that works for any number of bases
The old method did not generalize to an arbitrary number of bases.
Change to a new pair of rules -- equivalent to the old ones -- that
reason in terms of "file states in trees" (as in tree:file), as
opposed to tree diffs.  These rules generalize because we can express
the crucial "A:f agrees with _any_ Yi:f" etc.